feat: PC 랜딩 세종대왕 카드 스켈레톤 적용
