+++ src/constants/kingSejong.ts
@@ -1,0 +1,3 @@
+export const KING = {
+  uuid: 'dfjkasldfjkf0989fjkdf',
+};
